feat!: Cleanup qubit ops definitions, and improve the docs
